# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/checkout/ticket": {
      "post": {
        "operationId": "CheckoutController_ticket",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckoutTicketDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        }
      }
    },
    "/api/checkout/series-pass": {
      "post": {
        "operationId": "CheckoutController_seriesPass",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckoutSeriesPassDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        }
      }
    },
...
    "/api/tickets/{id}/transfer": {
      "post": {
        "operationId": "TicketsController_transfer",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransferTicketDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
...
    "/api/crm/people/tags": {
      "post": {
        "operationId": "CrmController_addTag",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTagDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        }
      }
    },
    "/api/crm/tags/{id}": {
      "delete": {
        "operationId": "CrmController_deleteTag",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/api/crm/scores/recompute": {
      "post": {
        "operationId": "CrmController_recompute",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActorRefDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/api/crm/campaigns": {
      "post": {
        "operationId": "CrmController_createCampaign",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCampaignDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        }
      },
      "get": {
        "operationId": "CrmController_listCampaigns",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/api/crm/campaigns/{id}/send": {
      "post": {
        "operationId": "CrmController_sendCampaign",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/api/crm/triggers": {
      "get": {
        "operationId": "CrmController_listTriggers",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        }
      },
      "post": {
        "operationId": "CrmController_createTrigger",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTriggerDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        }
      }
    },
    "/api/crm/triggers/{id}": {
      "patch": {
        "operationId": "CrmController_patchTrigger",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchTriggerDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/api/crm/triggers/evaluate": {
      "post": {
        "operationId": "CrmController_evaluateTriggers",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActorRefDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    }
  },
  "info": {
    "title": "omni-dance API",
    "description": "API de la plataforma omni-dance (SBK Santiago): eventos, QR, sesiones, tickets, social, academias, gamificación, RBAC y admin.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "omnidance_session"
      }
    },
    "schemas": {
...
      "CheckoutTicketDto": {
        "type": "object",
        "properties": {}
      },
      "CheckoutSeriesPassDto": {
        "type": "object",
        "properties": {}
      },
      "TransferTicketDto": {
        "type": "object",
        "properties": {}
      },
...
      "CreateTagDto": {
        "type": "object",
        "properties": {}
      },
      "ActorRefDto": {
        "type": "object",
        "properties": {}
      },
      "CreateCampaignDto": {
        "type": "object",
        "properties": {}
      },
      "CreateTriggerDto": {
        "type": "object",
        "properties": {}
      },
      "PatchTriggerDto": {
        "type": "object",
        "properties": {}
      }
    }
  }
}